refactor(chat): split 2840-LOC chat.py into focused submodules
chat.py was a single 2840-line file holding the 470-line system prompt,
the ChatSession streaming class, the run_chat dispatcher with 14 inline
slash-command handlers, and the run_demo scripted-demo function. Hard
to navigate; one logical change had to scroll past several others.

Extracted into:
- chat_prompt.py — SYSTEM_PROMPT (479 LOC)
- chat_session.py — ChatSession + provider-specific streaming (149 LOC)
- chat_demo.py — run_demo and its scene helpers (562 LOC)
- chat_commands.py — 14 handle_* slash-command handlers (1010 LOC)

chat.py is now 827 LOC: imports, run_chat input loop, prompt-toolkit
input box rendering, dispatch, LLM-response post-processing, main().
All public symbols remain importable from evalview.chat for backward
compat (dogfood/agent.py still imports ChatSession from there).

Side effect: imported `subprocess` at module top in chat.py. It was
previously imported lazily inside the /skill handler — now extracted —
which left a latent NameError on the LLM-response-execute path that
had never been exercised by tests but would have triggered at runtime.

Zero functional change otherwise. 1792 tests pass (same baseline).
